+++ Grocemate/Enums/HapticType.swift
@@ -1,0 +1,12 @@
+//
+//  HapticType.swift
+//  Grocemate
+//
+//  Created by Giorgio Latour on 11/23/23.
+//
+
+import Foundation
+
+public enum HapticType {
+    case swipeSuccess, longPressSuccess
+}
